Made forcing tcp upgrade for node local dns configurable. (#5034)
In certain scenarios, it might be beneficial to be able to keep upstream dns resolutions
on udp instead of forcing tcp as was the default for node local dns. For example, this
is true if the tcp connection to the upstream dns server is blocked via firewall or it
is simply reacting very slow to tcp requests. The default stays as is, though, i.e. forcing
tcp upgrades for upstream dns requests in the node local dns setup.